requirements.txt
numpy==1.21.5
pandas==1.4.2
matplotlib == 3.5.1
scikit-learn==1.0.2
joblib == 1.1.0
tpot == 0.11.7
pillow==9.0.1
protobuf==3.19.1
seaborn==0.11.2
click<=8.0.4
streamlit